Moved Store.js to lib directory
